2.ComputerVision/1.cv2/3.opencv_streamlit_intermediate/requirements.txt
streamlit
opencv-python
opencv-contrib-python
numpy
Pillow
pyzbar
imutils